backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.2.1
openai==2.14.0
pydantic==2.12.5
datamodel-code-generator==0.52.2
httpx==0.27.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
requests==2.31.0
faiss-cpu==1.8.0
numpy>=1.24.0
